⚡️ splits PageContent to own component to prevent full page re-render
